🐛 Fix[#26]: BuildingInfo-likeNum 수정
